build: refactor step organization
