+++ packages/scripts/src/steps/babel/default-config.ts
