+++ packages/scripts/src/steps/babel/index.ts
@@ -1,0 +1,1 @@
+export * from './babel'
