+++ packages/scripts/src/steps/index.ts
@@ -1,4 +1,3 @@
 export * from './noop'
-export * from './typedoc'
 export * from './typescript'
 export * from './babel'
