+++ packages/scripts/src/steps/typescript/index.ts
@@ -1,0 +1,3 @@
+export * from './compile'
+export * from './emitTypings'
+export * from './typedoc'
